# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de manipulação de produtos",
        "description": "Essa aplicação tem como objetivo a manipulação de produtos",
        "termsOfService": "http://localhost:3000/terms",
        "contact": {
            "email": "swagger@gmail.com"
        }
    },
    "version": "1.0.0",
    "servers": [
        {
            "url": "http://localhost:3000/v1",
            "description": "API de teste"
        }
    ],
    "paths": {
        "/products": {
            "post": {
                "sumary": "Cadastro de produtos",
                "description": "Essa rota será responsável na criação de um novo produto",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            },
                            "examples": {
                                "product": {
                                    "value": {
                                        "name": "Mouse USB",
                                        "description": "Produdo com ótima qualidade",
                                        "price": 24.20
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Product already exists"
                    },
                    "200": {
                        "description": "Ok",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "type": "Object",
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/findByname": {
            "get": {
                "description": "Busca pelo nome do produto",
                "summary": "Busca de um produto pelo nome",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nome do produto para busca",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Busca pelo id do produto",
                "summary": "Busca de um produto pelo id",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do produto para busca",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "type": "Object",
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Alterar produto pelo id",
                "summary": "Alterar produto pelo id",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do produto para busca",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            },
                            "examples": {
                                "product": {
                                    "value": {
                                        "name": "Mouse Cabo",
                                        "description": "Produdo com ótima qualidade",
                                        "price": 74.20
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "type": "Object",
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "id": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}